return gulong instead of guint.
Thu Jan 24 07:39:56 2002  Tim Janik  <timj@gtk.org>

        * gobject.c (g_signal_connect_object): return gulong instead
        of guint.

+++ gobject/gobject.h
@@ -198,7 +198,7 @@ void        g_value_set_object                (GValue         *value,
 					       gpointer        v_object);
 gpointer    g_value_get_object                (const GValue   *value);
 GObject*    g_value_dup_object                (const GValue   *value);
-guint	    g_signal_connect_object           (gpointer	       instance,
+gulong	    g_signal_connect_object           (gpointer	       instance,
 					       const gchar    *detailed_signal,
 					       GCallback       c_handler,
 					       gpointer	       gobject,
